+++ drizzlepac/acsData.py
@@ -6,8 +6,6 @@ Class used to model ACS specific instrument data.
 :License: :doc:`LICENSE`
 
 """
-from __future__ import absolute_import, division, print_function # confidence high
-
 from stsci.tools import fileutil
 import numpy as np
 from .imageObject import imageObject, baseImageObject