+++ drizzlepac/createMedian.py
@@ -6,11 +6,6 @@ Create a median image from the singly drizzled images.
 :License: :doc:`LICENSE`
 
 """
-
-# Import external packages
-from __future__ import (absolute_import, division, unicode_literals,
-                        print_function)
-
 import os
 import sys
 import math