+++ drizzlepac/imgclasses.py
@@ -7,7 +7,6 @@ Used by ``TweakReg``.
 :License: :doc:`LICENSE`
 
 """
-from __future__ import absolute_import, division, print_function
 import os
 import sys
 import copy